+++ config/default.js
@@ -3,12 +3,40 @@ const path = require('path');
 module.exports = {
   port: parseInt(process.env.PORT, 10) || 8565,
   persistor: {
-    type: 'file', // availabe: file, memory, s3
+    /**
+     * Available types of persistors:
+     *
+     * file:     Save images on local disk. On by default is the easiest to configure
+     * memory:   Save images in memory. Useful on development, can drain a lot of memory. Do not use
+     *           on production
+     * none: Never save the images. The images will be re-calculated each time. Very useful if
+     *           you plan to put a cache on front of this service, like CloudFront or CloudFlare.
+     * s3:       Save the images on Amazon S3
+     */
+    type: 'file',
     options: {
       basePath: path.resolve(__dirname, '../cache'),
     },
 
     /**
+     * File
+     * type: 'file',
+     * options: {
+     *  basPath: '/tmp/',                 // The folder to store the images in.
+     * }
+     *
+     *
+     * Memory
+     * type: 'memory',
+     * options: {
+     *  limit: 100,                       // The limit expressed in MB. Can be false for no limit
+     * }
+     *
+     * Disk-less
+     * type: 'none',
+     * options: {}
+     *
+     *
      * S3 persistor
      *
      * type: 's3',
@@ -20,6 +48,12 @@ module.exports = {
      * }
     */
   },
+  /**
+   * Allow only files in the domain whitelist.
+   * This can be an array of domains and wildcard are supported
+   * whitelist: ["images.google.com", "*.mydomain.net"]
+   */
+  whitelist: false,
   compress: true,
 };