+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+
+- Support new persistors
+  - memory, none
+- Error handler
+- Domain whitelist
+- Docker version (initial)
 ## 1.0.1
 
 - Basic working service